Refactoring __init__ for rbsp. Only a few tests have been performed.
